refactor: move the lib folder from app to src
